chore: AnchorView 코드 이동
#77
